moved the cli
